get quorum fields to match, fixing test
Writing this commit message in anger:

It seems that the HBaseSource must match exactly between the
instantiated Job class and the JobTest. The error when this isn't the
case is very obscure: a `None.get()` exception deep in SpyGlass
internals. Blech.

This may or may not explain other test failure issues.