GroupFatcatWorksSubsetJob
This is a hack-y variant of GroupFatcatWorksSubsetJob which allows
setting different left and right sides of the join. The initial
application is to re-run work merging with only longtail-oa works on the
"left", with the goal of hard-merging these releases into existing
releases with actual identifiers (instead of just grouping into works).

As a refactor, the normal GroupFatcatWorksJob could just be this with
the same file passed as both left and right, though that requires twice
as much JSON parsing/filtering.