grobid persist: if status_code is not set, default to 0
We have to set something currently because of a NOT NULL constraint on
the table.

Originally I thought we would just not record rows if there was an
error, and that is still sort of a valid stance. However, when doing
bulk GROBID-ing from cdx table, there exist some "bad" CDX rows which
cause wayback or petabox errors. We should fix bugs or delete these rows
as a cleanup, but until that happens we should record the error state so
we don't loop forever.

One danger of this commit is that we can clobber existing good rows with
new errors rapidly if there is wayback downtime or something like that.