hack-y backoff ingest attempt
The goal here is to have SPNv2 requests backoff when we get
back-pressure (usually caused by some sessions taking too long). Lack of
proper back-pressure is making it hard to turn up parallelism.

This is a hack because we still timeout and drop the slow request. A
better way is probably to have a background thread run, while the
KafkaPusher thread does polling. Maybe with timeouts to detect slow
processing (greater than 30 seconds?) and only pause/resume in that
case. This would also make taking batches easier. Unlike the existing
code, however, the parallelism needs to happen at the Pusher level to do
the polling (Kafka) and "await" (for all worker threads to complete)
correctly.