persist: only GROBID updates file_meta, not file-result
The hope here is to reduce deadlocks in production (on aitio).

As context, we are only doing "updates" until the entire file_meta table
is filled in with full metadata anyways; updates are wasteful of
resources, and most inserts we have seen the file before, so should be
doing "DO NOTHING" if the SHA1 is already in the table.