SPNv2: make 'resources' optional
This was always present previously. A change was made to SPNv2 API
recently that borked it a bit, though in theory should be present on new
captures. I'm not seeing it for some captures, so pushing this work
around. It seems like we don't actually use this field anyways, at least
for ingest pipeline.