file ingest: don't 'backoff' on spn2 backoff error
The intent of this is to try and get through the daily ingest requests
faster, so we can loop and retry if needed. A 200 second delay, usually
resulting in a kafka topic reshuffle, really slows things down. This
will presumably result in a bunch of spn2-backoff status requests, but
we can just retry those.